% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_best_model_adj_pvalue}
\alias{add_best_model_adj_pvalue}
\title{Add Best Model Adjusted P-value to DoseRider Object}
\usage{
add_best_model_adj_pvalue(doseRiderObj)
}
\arguments{
\item{doseRiderObj}{A DoseRider object containing results from dose-response analysis.}
}
\value{
A DoseRider object with an added column "best_model_adj_pvalue".
}
\description{
This function adds a column "best_model_adj_pvalue" to a DoseRider object, containing the adjusted p-value
corresponding to the best model for each gene set.
}
\examples{
\dontrun{
updated_results <- add_best_model_adj_pvalue(doseRiderObj)
}

}
