% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting_helpers.R
\name{add_cluster_trends_and_bmd}
\alias{add_cluster_trends_and_bmd}
\title{Add cluster trends and BMD lines to the plot}
\usage{
add_cluster_trends_and_bmd(
  p,
  gene_set_results,
  mean_data,
  dose_col,
  draw_bmd,
  v_size,
  clusterResults,
  amount
)
}
\description{
Add cluster trends and BMD lines to the plot
}
