% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting_helpers.R
\name{add_gene_annotations}
\alias{add_gene_annotations}
\title{Add gene annotations to the plot}
\usage{
add_gene_annotations(p, mean_data, dose_col, annotation_text_size)
}
\description{
Add gene annotations to the plot
}
