% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_geneset.R
\name{apply_pathway_filters}
\alias{apply_pathway_filters}
\title{Generalized Pathway Filtering Function}
\usage{
apply_pathway_filters(
  long_df,
  filters = c("pca", "variance", "correlation"),
  filter_params = list()
)
}
\arguments{
\item{long_df}{Dataframe in long format with transcriptomic data.}

\item{filters}{List of filtering methods to apply. Options: "pca", "variance", "correlation".}

\item{filter_params}{Named list containing parameters for each filtering method.}
}
\value{
Logical value: TRUE if the pathway passes all selected filters, FALSE otherwise.
}
\description{
Generalized Pathway Filtering Function
}
