% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseRider.R
\name{as.data.frame.DoseRider}
\alias{as.data.frame.DoseRider}
\title{Convert a DoseRider object to a data frame}
\usage{
\method{as.data.frame}{DoseRider}(object)
}
\arguments{
\item{object}{A DoseRider object.}
}
\value{
A data frame with specific attributes from the DoseRider object.
Each row of the data frame corresponds to an element of the DoseRider object.
}
\description{
This function extracts specific attributes from a DoseRider object and
structures them in a data frame for easier manipulation and visualization.
The attributes included are: Geneset, Geneset_Size, GeneRatio, bgRatio,
Genes, Base_AIC, Base_BIC, Base_edf, Linear_AIC, Linear_BIC, Linear_edf,
Cubic_AIC, Cubic_BIC, Cubic_edf, P_Value_Linear, P_Value_Cubic, and Adjusted_P_Value.
}
\examples{
\dontrun{
  # Assuming `dose_rider_result` is a DoseRider object
  result_df <- as.data.frame.DoseRider(dose_rider_result)
}

}
