% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_original_tcds}
\alias{cluster_original_tcds}
\title{Cluster Original TCD Values}
\usage{
cluster_original_tcds(original_tcds, method = "kmeans")
}
\arguments{
\item{original_tcds}{A numeric vector of original TCD values.}

\item{method}{Clustering method. Options are "kmeans" (default) or "gmm" (Gaussian Mixture Models).}
}
\value{
A list containing the number of clusters (`k`) and the clustering results.
}
\description{
This function clusters the original TCD values and determines the number of clusters to use for bootstrap data.
If the number of unique TCD values is less than 3, it assigns the number of clusters as the length of unique TCDs
and returns a named vector for clustering.
}
