% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compareGAMM}
\alias{compareGAMM}
\title{Compare Two GAM Models}
\usage{
compareGAMM(model1, model2)
}
\arguments{
\item{model1}{A gam or bam model object. The first model to compare.}

\item{model2}{A gam or bam model object. The second model to compare.}
}
\value{
The p-value from the Chi-square test comparing the two models.
}
\description{
This function compares two Generalized Additive Models (GAMs) by performing a Chi-square test on their scores.
The function first checks if both models are GAMs and if they are built using the same method.
It then computes the test statistic based on the difference of the models' Un-Biased Risk Estimators (UBREs)
and their degrees of freedom.
}
\examples{
data(mtcars)
fit1 <- mgcv::gam(mpg ~ s(hp), data = mtcars)
fit2 <- mgcv::gam(mpg ~ s(hp, k = 5), data = mtcars)
compareGAMM(fit1, fit2)

}
