% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd.R
\name{compute_bmd_bounds}
\alias{compute_bmd_bounds}
\title{Compute BMD Bounds using Bootstrapping}
\usage{
compute_bmd_bounds(
  dose_rider_results,
  dose_col = "dose",
  sample_col = "sample",
  covariates = c(),
  omic = "rnaseq",
  n_bootstrap = 100,
  ci_level = 0.95,
  filter_type = "fdr",
  threshold = 0.1,
  top = 10,
  model_type = "all"
)
}
\arguments{
\item{dose_rider_results}{A DoseRider object containing results of the analysis.}

\item{dose_col}{Name of the column representing dose information.}

\item{sample_col}{Name of the column representing sample information.}

\item{covariates}{Optional, vector specifying the covariate column(s) in `se`.}

\item{omic}{Type of omics data, defaults to "rnaseq".}

\item{n_bootstrap}{The number of bootstrap samples to generate. Default is 1000.}

\item{ci_level}{Confidence interval level, default is 0.95.}

\item{filter_type}{The type of p-value to filter by. Options are "pvalue" for raw p-value and "fdr" for adjusted p-value.}

\item{threshold}{The p-value threshold for filtering. Defaults to 0.05.}

\item{model_type}{The type of model to filter by. Options are "all", "linear", "non_linear" (which includes "non_linear_fixed" and "non_linear_mixed").}
}
\value{
A list containing the lower and upper BMD bounds for each significant gene set.
}
\description{
This function calculates the lower and upper bounds of the Benchmark Dose (BMD)
using bootstrapping on the raw data for significant gene sets.
}
\examples{
\dontrun{
data("SummarizedExperiment")
gmt <- list(geneSet1 = list(genes = c("gene1", "gene2")))
results <- DoseRider(se, gmt, "dose", "sample", "covariate", "rnaseq", modelType = "GAMM")
bmd_bounds <- compute_bmd_bounds(results, "dose", "sample", n_bootstrap = 1000, ci_level = 0.95)
print(bmd_bounds)
}

}
