% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd.R
\name{compute_bmd_bounds_sequential}
\alias{compute_bmd_bounds_sequential}
\title{Compute BMD Bounds Sequentially}
\usage{
compute_bmd_bounds_sequential(
  dose_rider_results,
  dose_col,
  sample_col,
  ci_level = 0.95,
  covariates = c(),
  omic = "rnaseq",
  n_bootstrap = 1000,
  clusterResults = FALSE,
  z = 1
)
}
\arguments{
\item{dose_rider_results}{A DoseRider object containing results of the analysis.}

\item{dose_col}{Name of the column representing dose information.}

\item{sample_col}{Name of the column representing sample information.}

\item{ci_level}{Confidence interval level, default is 0.95.}

\item{covariates}{Vector specifying the covariate column(s).}

\item{omic}{Type of omics data, defaults to "rnaseq".}

\item{n_bootstrap}{The number of bootstrap samples to generate.}

\item{clusterResults}{Logical, whether to compute cluster-specific results.}

\item{z}{A parameter passed to the model fitting function.}
}
\value{
A data frame with BMD bounds and TCD cluster statistics for all gene sets.
}
\description{
Compute BMD Bounds Sequentially
}
