% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd.R
\name{compute_bmd_from_main_trend}
\alias{compute_bmd_from_main_trend}
\title{Compute Benchmark Dose (BMD) from Main Trend}
\usage{
compute_bmd_from_main_trend(
  model,
  data,
  dose_var,
  omic,
  z = 1,
  covariates = c()
)
}
\arguments{
\item{model}{A fitted model object, either of class `lmerMod` or `glmerMod`.}

\item{data}{A dataframe containing the variables used in the model.}

\item{dose_var}{The name of the dose variable in `data`.}

\item{omic}{A character string indicating the type of omics data.
Use "rnaseq" for RNA-Seq data; otherwise, specify another type.}

\item{z}{The number of standard deviations away from the control response
to define the target response.}
}
\value{
A dataframe containing the Benchmark Dose (BMD) for each specified level of `z`.
}
\description{
This function calculates the Benchmark Dose (BMD) for a given model and dataset.
BMD is determined as the dose at which the predicted response is a certain
number of standard deviations (defined by `z`) away from the control response.
The function is designed to work with models fitted using `lmer` or `glmer`
from the `lme4` package and is applicable to linear and generalized linear mixed models.
}
\examples{
# Assuming `model` is a fitted lmerMod or glmerMod object
# `data` is a dataframe with dose, size_factors, and other covariates
compute_bmd_from_main_trend(model, data, "dose", "rnaseq", z = 1)
}
