% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamm.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute metrics for a given model}
\usage{
compute_metrics(model)
}
\arguments{
\item{model}{A model object or a list. If it is a model, the function computes the metrics. If it's not a list, the function returns NA for all the metrics.}
}
\value{
A list containing the AIC, BIC and edf of the model. If the model parameter is not a list, the function will return NA for all metrics.
}
\description{
This function computes the AIC, BIC and edf metrics for a given model. If the model is not a list, it will return NA for all the metrics.
}
\examples{
# Create a linear model
data(mtcars)
model <- lm(mpg ~ hp, data = mtcars)

# Compute metrics
compute_metrics(model)
}
