% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamm.R
\name{compute_metrics_gamm}
\alias{compute_metrics_gamm}
\title{Compute metrics for a given gam or bam model}
\usage{
compute_metrics_gamm(model)
}
\arguments{
\item{model}{A `gam` or `bam` model object from the mgcv package.}
}
\value{
A list containing the AIC, BIC, and edf of the model.
        If the model parameter isn't a `gam` or `bam` object, the function will return NA for all metrics.
}
\description{
This function computes the AIC, BIC, and effective degrees of freedom (edf) for a given `gam` or `bam` model.
}
\details{
The approach to calculate the effective degrees of freedom (edf)
is adapted from the itsadug package: Interpreting Time Series and Autocorrelated Data Using GAMMs
Reference: itsadug package (https://CRAN.R-project.org/package=itsadug)
}
\examples{
library(mgcv)
data(sleepstudy)
gam_model <- gam(Reaction ~ s(Days), data = sleepstudy)

# Compute metrics
compute_metrics_gamm(gam_model)
}
