% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{compute_metrics_lmm}
\alias{compute_metrics_lmm}
\title{Compute metrics for a given lmer or glmer model}
\usage{
compute_metrics_lmm(model)
}
\arguments{
\item{model}{A `lmer` or `glmer` model object. If it's neither, the function returns NA for all the metrics.}
}
\value{
A list containing the AIC, BIC, and degrees of freedom of the model. If the model parameter isn't a `lmer` or `glmer` object, the function will return NA for all metrics.
}
\description{
This function computes the AIC, AICc, BIC, and degrees of freedom metrics for a given `lmer` or `glmer` model.
}
\examples{
library(lme4)
data(sleepstudy)
model <- lmer(Reaction ~ Days + (1|Subject), data = sleepstudy)

# Compute metrics
compute_metrics(model)
}
