% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{compute_tcd_statistics}
\alias{compute_tcd_statistics}
\title{Compute TCD Statistics}
\usage{
compute_tcd_statistics(tcd_values, ci_level = 0.95)
}
\arguments{
\item{ci_level}{Confidence interval level (default is 0.95).}

\item{bmd_values}{A numeric vector of BMD values.}
}
\value{
A named list containing `lower_bound`, `upper_bound`, `mean_bmd`, and `median_bmd`.
}
\description{
This function calculates key statistics for BMD values, including lower and upper bounds,
mean, and median, based on a given confidence interval level.
}
