% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CPDB_pathways_genes_entrez}
\alias{CPDB_pathways_genes_entrez}
\title{#' CPDB Pathways and Genes (ENSEMBL)
#'
#' A dataset containing the ConsensusPathDB (CPDB) pathways and their associated genes
#' identified by ENSEMBL gene IDs.
#'
#' @format A data frame with the following columns:
#' \describe{
#'   \item{Pathway}{A character vector with the name of the pathway.}
#'   \item{Gene}{A character vector with the ENSEMBL gene IDs associated with the pathway.}
#' }
#' @source \url{http://consensuspathdb.org/}
"CPDB_pathways_genes_ensembl"
CPDB Pathways and Genes (Entrez)}
\format{
A data frame with the following columns:
\describe{
  \item{Pathway}{A character vector with the name of the pathway.}
  \item{Gene}{A character vector with the Entrez gene IDs associated with the pathway.}
}
}
\source{
\url{http://consensuspathdb.org/}
}
\usage{
CPDB_pathways_genes_entrez
}
\description{
A dataset containing the ConsensusPathDB (CPDB) pathways and their associated genes
identified by Entrez gene IDs.
}
\keyword{datasets}
