% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CPDB_pathways_genes_refseq}
\alias{CPDB_pathways_genes_refseq}
\title{CPDB Pathways and Genes (RefSeq)}
\format{
A data frame with the following columns:
\describe{
  \item{Pathway}{A character vector with the name of the pathway.}
  \item{Gene}{A character vector with the RefSeq gene IDs associated with the pathway.}
}
}
\source{
\url{http://consensuspathdb.org/}
}
\usage{
CPDB_pathways_genes_refseq
}
\description{
A dataset containing the ConsensusPathDB (CPDB) pathways and their associated genes
identified by RefSeq gene IDs.
}
\keyword{datasets}
