% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CPDB_pathways_metabolites}
\alias{CPDB_pathways_metabolites}
\title{CPDB Pathways and Metabolites}
\format{
A data frame with the following columns:
\describe{
  \item{Pathway}{A character vector with the name of the pathway.}
  \item{Metabolite}{A character vector with the identifiers of metabolites associated with the pathway.}
}
}
\source{
\url{http://consensuspathdb.org/}
}
\usage{
CPDB_pathways_metabolites
}
\description{
A dataset containing the ConsensusPathDB (CPDB) pathways and their associated metabolites.
}
\keyword{datasets}
