% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamm.R
\name{create_gamm_formula}
\alias{create_gamm_formula}
\title{Create Generalized Additive Mixed Model (GAMM) Formula}
\usage{
create_gamm_formula(
  response,
  fixed_effects,
  random_effects,
  covariates = c(),
  model_type = "base"
)
}
\arguments{
\item{response}{A string specifying the response variable in the model.}

\item{fixed_effects}{A string or a vector of strings specifying the fixed effects in the model.}

\item{random_effects}{A string or a vector of strings specifying the random effects in the model.}

\item{covariates}{A string or a vector of strings specifying any covariates to be included in the model. Default is an empty vector.}

\item{model_type}{A string specifying the type of the model. Can be 'base', 'linear', or 'cubic'. Default is 'base'.}
}
\value{
A string representing the formula for the GAMM.
}
\description{
This function generates a formula for a GAMM, given response and effect variables, and the model type.
It allows for the specification of 'base', 'linear', or 'cubic' model types. The main goal is to
compare the 'cubic' model with the 'base' model to evaluate if the fixed effect (in this case 'dose')
is significant.
}
\examples{
\dontrun{
# Create a base formula
base_formula <- create_gamm_formula(response = "counts",
                                    fixed_effects = "dose",
                                    random_effects = "gene",
                                    model_type = "base")
# Create a linear formula
linear_formula <- create_gamm_formula(response = "counts",
                                      fixed_effects = "dose",
                                      random_effects = "gene",
                                      model_type = "linear")
# Create a cubic formula
cubic_formula <- create_gamm_formula(response = "counts",
                                     fixed_effects = "dose",
                                     random_effects = "gene",
                                     model_type = "cubic")
}
}
