% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{create_gene_heatmap}
\alias{create_gene_heatmap}
\title{Create Heatmap for Individual Genes within a Gene Set}
\usage{
create_gene_heatmap(
  dose_rider_results,
  gene_set_name,
  dose_col,
  dose_unit = "μM",
  fontsize = 6
)
}
\arguments{
\item{dose_rider_results}{A list containing the results of the DoseRider analysis for each gene set.}

\item{gene_set_name}{A character string specifying the name of the gene set to be visualized.}

\item{dose_col}{A character string specifying the name of the dose column in the raw expression data.}

\item{dose_unit}{A character string specifying the dose units to plot in the column title..}

\item{fontsize}{Integer for fontsize. Default 6}
}
\value{
An object of class `Heatmap` representing the constructed heatmap for the specified gene set.
}
\description{
This function creates a heatmap for a specified gene set from DoseRider results. Each row in the heatmap
represents an individual gene, and each column represents different doses, showing gene expression values.
}
\examples{
\dontrun{
  # Assuming `dose_rider_results` is your DoseRider analysis result
  heatmap_plot <- create_gene_heatmap(dose_rider_results, "Gene Set Name", "dose")
  draw(heatmap_plot) # To display the heatmap
}

}
