% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseRider.R
\name{DoseRider}
\alias{DoseRider}
\title{DoseRider Function}
\usage{
DoseRider(
  se,
  gmt,
  dose_col = "dose",
  sample_col = "sample",
  covariate = "",
  omic = "rnaseq",
  minGSsize = 5,
  maxGSsize = 300,
  method = "fdr"
)
}
\arguments{
\item{se}{The gene expression data as either a SummarizedExperiment object or a matrix/data frame.}

\item{gmt}{A list of gene sets, each gene set is a list with a 'genes' element containing a character vector of gene symbols.}

\item{dose_col}{The name of the dose column in the metadata.}

\item{sample_col}{The name of the sample column in the metadata.}

\item{covariate}{(Optional) The name of a covariate column in the metadata.}

\item{omic}{The omic type, defaults to "rnaseq".}

\item{minGSsize}{Minimum gene set size for consideration, defaults to 5.}

\item{maxGSsize}{Maximum gene set size for consideration, defaults to 300.}

\item{method}{The method for multiple testing adjustment, defaults to "fdr".}

\item{metadata}{(Optional) A dataframe that holds metadata information. This is required when 'se' is a matrix or data frame.}
}
\value{
A list with results from the dose-response analysis for each gene set.
Each list contains the following items:
- `Geneset`: The name of the gene set.
- `Geneset_Size`: The number of genes in the gene set.
- `GeneRatio`: The proportion of genes from the gene set present in the data.
- `bgRatio`: The proportion of genes in the background set.
- `Genes`: The number of unique genes in the long_df.
- `Base_AIC`: Akaike Information Criterion for the base model.
- `Base_BIC`: Bayesian Information Criterion for the base model.
- `Base_edf`: Effective degrees of freedom for the base model.
- `Linear_AIC`: Akaike Information Criterion for the linear model.
- `Linear_BIC`: Bayesian Information Criterion for the linear model.
- `Linear_edf`: Effective degrees of freedom for the linear model.
- `Cubic_AIC`: Akaike Information Criterion for the cubic model.
- `Cubic_BIC`: Bayesian Information Criterion for the cubic model.
- `Cubic_edf`: Effective degrees of freedom for the cubic model.
- `P_Value_Linear`: P-value for the linear model.
- `P_Value_Cubic`: P-value for the cubic model.
- `Smooth_Predictions`: A list containing smoothed predictions from the cubic model.
- `Adjusted_P_Value`: P-value for the cubic model adjusted for multiple testing.
}
\description{
This function performs a series of analysis on gene expression data
provided as a SummarizedExperiment object or a matrix/data frame with metadata.
It uses Generalized Additive Mixed Models (GAMMs) to analyze the relationship
between dose-response and gene sets.
}
