% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd.R
\name{extract_bmd_for_pathway}
\alias{extract_bmd_for_pathway}
\title{Extract Benchmark Dose (BMD) Values for a Pathway}
\usage{
extract_bmd_for_pathway(dose_rider_results, gene_set_name)
}
\arguments{
\item{dose_rider_results}{A list containing the results of DoseRider analysis.}

\item{gene_set_name}{The name of the gene set/pathway.}
}
\value{
A numeric vector containing the BMD values for the specified pathway. If no cluster-specific results are available, NA is returned.
}
\description{
This function extracts the Benchmark Dose (BMD) values for all clusters within a specified pathway from the DoseRider results.
}
\examples{
\dontrun{
dose_rider_results <- DoseRider(se, gmt, "dose", "sample")
bmd_values <- extract_bmd_for_pathway(dose_rider_results, "geneSet1")
print(bmd_values)
}

}
