% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting_helpers.R
\name{extract_gene_set_results}
\alias{extract_gene_set_results}
\title{Extract gene set results from dose_rider_results}
\usage{
extract_gene_set_results(dose_rider_results, gene_set_name, plot_original_data)
}
\description{
Extract gene set results from dose_rider_results
}
