% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamm.R
\name{extract_random_effects_gamm}
\alias{extract_random_effects_gamm}
\title{Extract Random Effects from GAMM}
\usage{
extract_random_effects_gamm(model, dose_col)
}
\arguments{
\item{model}{A GAMM model object.}
}
\value{
A dataframe with genes as rows and their random effect estimates.
}
\description{
This function extracts the random effect estimates for each gene from a GAMM object.
The result is a dataframe with genes as rows and their corresponding random effect values.
}
