% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{extract_random_effects_lmm}
\alias{extract_random_effects_lmm}
\title{Extract Random Effects from LMM}
\usage{
extract_random_effects_lmm(model, dose_col)
}
\arguments{
\item{model}{An LMM model object, typically of class 'lmerMod'.}
}
\value{
A dataframe with genes as rows and their random effect estimates.
}
\description{
This function extracts the random effect estimates for each gene from an LMM model object.
The result is a dataframe with genes as rows and their corresponding random effect values.
}
