% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd.R
\name{extract_tcd_for_pathway}
\alias{extract_tcd_for_pathway}
\title{Extract Minimum TCD1 Value for a Pathway}
\usage{
extract_tcd_for_pathway(dose_rider_results, gene_set_name)
}
\arguments{
\item{dose_rider_results}{A list containing the results of DoseRider analysis.}

\item{gene_set_name}{The name of the gene set/pathway.}
}
\value{
A numeric value representing the minimum TCD1 for the specified pathway.
        If no cluster-specific results are available, NA is returned.
}
\description{
This function extracts the minimum TCD1 (Threshold Concentration Dose, second derivative zero points)
value for a specified pathway from the DoseRider results.
}
\examples{
\dontrun{
dose_rider_results <- DoseRider(se, gmt, "dose", "sample")
tcd1_value <- extract_tcd_for_pathway(dose_rider_results, "geneSet1")
print(tcd1_value)
}

}
