% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseRider.R
\name{filter_DoseRider}
\alias{filter_DoseRider}
\title{Filter DoseRider Object}
\usage{
filter_DoseRider(
  doseRiderObj,
  model_type = "all",
  filter_type = "pvalue",
  threshold = 0.05
)
}
\arguments{
\item{doseRiderObj}{A DoseRider object containing results from dose-response analysis.}

\item{model_type}{The type of model to filter by. Options are "all", "linear", "non_linear" (which includes "non_linear_fixed" and "non_linear_mixed").}

\item{filter_type}{The type of p-value to filter by. Options are "pvalue" for raw p-value and "fdr" for adjusted p-value.}

\item{threshold}{The p-value threshold for filtering. Defaults to 0.05.}
}
\value{
A filtered DoseRider object.
}
\description{
This function filters a DoseRider object based on the specified model type and p-value criteria.
}
\examples{
\dontrun{
filtered_results <- filter_DoseRider(doseRiderObj, model_type = "linear", filter_type = "fdr", threshold = 0.05)
}

}
