% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset.R
\name{filter_gmt_by_size}
\alias{filter_gmt_by_size}
\title{Filter GMT by Geneset Size}
\usage{
filter_gmt_by_size(gmt, minGenesetSize, maxGenesetSize)
}
\arguments{
\item{gmt}{A list, where each element represents a geneset. Each geneset should be a named list with a field for the geneset name and a field for the genes contained in the geneset.}

\item{minGenesetSize}{An integer specifying the minimum number of genes that a geneset must contain to be kept.}

\item{maxGenesetSize}{An integer specifying the maximum number of genes that a geneset can contain to be kept.}
}
\value{
A list (like the original GMT object) but only containing the genesets that met the size criteria.
}
\description{
This function filters a GMT (Gene Matrix Transposed file format) object by a specified geneset size range.
Each geneset in the GMT object is checked to see if its size falls within the minimum and maximum size thresholds.
If it does, the geneset is kept; if it doesn't, the geneset is discarded.
}
\examples{
# Consider a GMT object
gmt <- list(list(pathway = "geneset1", genes = c("gene1", "gene2", "gene3")),
             list(pathway = "geneset2", genes = c("gene4", "gene5")),
             list(pathway = "geneset3", genes = c("gene6", "gene7", "gene8", "gene9", "gene10")))

# Filter the GMT object
filter_gmt_by_size(gmt, minGenesetSize = 3, maxGenesetSize = 4)
# This will return the first geneset only, as it is the only one with size between 3 and 4

}
