% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset.R
\name{find_geneset_index}
\alias{find_geneset_index}
\title{Find the Index of a Geneset in a GMT Object}
\usage{
find_geneset_index(gmt, geneset_name, pathway_col)
}
\arguments{
\item{gmt}{A list, where each element represents a geneset. Each geneset should be a named list with at least a field corresponding to the geneset name.}

\item{geneset_name}{A character string representing the name of the geneset to find.}

\item{pathway_col}{A character string representing the name of the field in the geneset list items which corresponds to the geneset name.}
}
\value{
The index of the geneset in the GMT object if found, otherwise NULL.
}
\description{
This function searches for a specified geneset within a GMT (Gene Matrix Transposed file format) object and returns its index.
If the geneset is not found, NULL is returned.
}
\examples{
# Consider a GMT object
gmt <- list(list(pathway = "geneset1", gene_ids = c("gene1", "gene2")),
             list(pathway = "geneset2", gene_ids = c("gene3", "gene4")))

# Find the index of a geneset
find_geneset_index(gmt, "geneset2", "pathway") # returns 2
find_geneset_index(gmt, "nonexistent_geneset", "pathway") # returns NULL

}
