% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamm.R
\name{fit_gam}
\alias{fit_gam}
\title{Fit Generalized Additive Model (GAM) and Compute AIC and BIC}
\usage{
fit_gam(formula, data)
}
\arguments{
\item{formula}{A formula for the GAM model generated by the `create_gamm_formula` function.}

\item{data}{A data frame (in long format) containing the data to be modeled. The data frame should include
columns for dose, gene (or metabolite), and, if 'rnaseq' is specified, an offset. Other covariates may also be present.}
}
\value{
A GAM model if fitting is successful; NA otherwise.
}
\description{
This function fits a GAM to the provided data using the specified formula.
The aim is to model all the genes from a specified pathway to check the effects of a dose.
The model is fitted using the maximum likelihood (ML) method, as this allows for the comparison of models with random effects.
If fitting fails, NA is returned.
}
\examples{
\dontrun{
data("mtcars")
# Generate a suitable formula for the GAM model using `create_gamm_formula`
formula <- create_gamm_formula(response = "mpg",
                               fixed_effects = "hp",
                               random_effects = "cyl",
                               model_type = "base")
model <- fit_gam(formula = formula, data = mtcars)
}

}
