% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{fit_lmm}
\alias{fit_lmm}
\title{Fit Linear Mixed Model (LMM) or Generalized Linear Mixed Model (GLMM)}
\usage{
fit_lmm(formula, data, omic = "base")
}
\arguments{
\item{formula}{A formula for the LMM or GLMM.}

\item{data}{A data frame (in long format) containing the data to be modeled. The data frame should include
columns for dose, gene (or metabolite), and, if 'rnaseq' is specified, an offset. Other covariates may also be present.}

\item{omic}{A character string specifying the type of data. If 'rnaseq', the family is set to negative binomial.}

\item{theta}{The theta parameter for the negative binomial family. Relevant only if omic = "rnaseq".}
}
\value{
An LMM or GLMM model if fitting is successful; NA otherwise.
}
\description{
This function fits an LMM or a GLMM to the provided data using the specified formula.
The aim is to model all the genes from a specified pathway to check the effects of a dose.
The model's family is chosen based on the 'omic' parameter.
}
\examples{
\dontrun{
data("mtcars")
# Generate a suitable formula for the LMM
formula <- mpg ~ hp + (1|cyl)
model <- fit_lmm(formula = formula, data = mtcars, omic = "rnaseq", theta = 1.5)
}

}
