% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd.R
\name{fit_model_compute_bmd}
\alias{fit_model_compute_bmd}
\title{Fit Model and Compute BMD on Bootstrapped Data}
\usage{
fit_model_compute_bmd(
  long_df,
  formula,
  omic = "rnaseq",
  clusterResults = FALSE,
  dose_col,
  z = 1,
  keep_min_bmd = TRUE
)
}
\arguments{
\item{long_df}{Data frame containing the raw data for the gene set.}

\item{formula}{The formula used to fit the model (LMM or GAMM).}

\item{omic}{Type of omics data, defaults to "rnaseq".}

\item{clusterResults}{Boolean, if TRUE, compute BMD for clusters; otherwise, for the whole gene set. Default is FALSE.}

\item{dose_col}{Name of the column representing dose information.}
}
\value{
A list containing the BMD for the bootstrapped sample.
}
\description{
This function fits a model to a bootstrapped sample and computes the BMD,
either for the entire gene set or for clusters within the gene set.
}
\examples{
\dontrun{
long_df_bootstrap <- long_df[sample(nrow(long_df), replace = TRUE), ]
bootstrap_results <- fit_model_compute_bmd(long_df_bootstrap, formula, "rnaseq", clusterResults = FALSE, dose_col = "dose")
print(bootstrap_results)
}
}
