% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd.R
\name{get_bmd_range}
\alias{get_bmd_range}
\title{Plot Benchmark Dose (BMD) Density and Peaks}
\usage{
get_bmd_range(dose_rider_results)
}
\arguments{
\item{bmd_range_output}{A list containing the output from `get_bmd_range` function,
which includes x (BMD values), y (density), and bmd (peaks).}
}
\value{
A ggplot object visualizing the density of BMD values with peaks marked.
}
\description{
This function creates a plot visualizing the density of BMD values and highlights
the peaks where the highest density of BMD values are found.
}
\examples{
bmd_range_output <- get_bmd_range(dose_rider_results)
plot_bmd_density_and_peaks(bmd_range_output)

}
