% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd.R
\name{get_tcd_range}
\alias{get_tcd_range}
\title{Compute Minimum TCD1 Values and Find Zero Points in Density}
\usage{
get_tcd_range(dose_rider_results)
}
\arguments{
\item{dose_rider_results}{A list containing the results of DoseRider analysis.}
}
\value{
A list containing the TCD density (`x`, `y`) and the identified zero points (`tcd`).
}
\description{
This function computes the density of minimum TCD1 (Threshold Concentration Dose,
second derivative zero points) values from the DoseRider results and identifies
zero points in the density.
}
\examples{
\dontrun{
dose_rider_results <- DoseRider(se, gmt, "dose", "sample")
tcd_range <- get_tcd_range(dose_rider_results)
print(tcd_range)
}

}
