% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_top_genesets}
\alias{get_top_genesets}
\title{Function to obtain the top k gene sets sorted by a specified column}
\usage{
get_top_genesets(
  dose_rider_results,
  pvalue_column = "best_model_adj_pvalue",
  order_column = "NegLogPValue",
  top = 10,
  decreasing = TRUE
)
}
\arguments{
\item{dose_rider_results}{Data frame containing gene set information}

\item{pvalue_column}{Column name for p-values (default: "best_model_adj_pvalue")}

\item{order_column}{Column name for ordering (default: "NegLogPValue")}

\item{top}{Number of top gene sets to retrieve (default: 10)}

\item{decreasing}{Logical indicating sort order (default: TRUE)}
}
\value{
Vector of gene set names of the top k gene sets
}
\description{
Function to obtain the top k gene sets sorted by a specified column
}
