% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting_helpers.R
\name{initialize_plot}
\alias{initialize_plot}
\title{Initialize the ggplot object with gene-specific trends}
\usage{
initialize_plot(mean_data, dose_col, model_metrics, gene_set_name)
}
\description{
Initialize the ggplot object with gene-specific trends
}
