% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset.R
\name{loadCPDB}
\alias{loadCPDB}
\title{Load CPDB GMT File}
\usage{
loadCPDB(identifier)
}
\arguments{
\item{identifier}{A character string specifying the desired identifier.
Options include 'ENSEMBL', 'Entrez', 'RefSeq', 'Symbol', 'Metabolites'.}
}
\value{
The loaded GMT file as a data frame.
}
\description{
This function loads the GMT file corresponding to the desired identifier from the ConsensusPathDB.
}
\examples{
loadCPDB("ENSEMBL")
}
