% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{optimal_clusters_silhouette}
\alias{optimal_clusters_silhouette}
\title{Determine Optimal Number of Clusters Using Silhouette Method}
\usage{
optimal_clusters_silhouette(data, dose_col, max_clusters = 10)
}
\arguments{
\item{data}{Dataframe containing dose-response data.}

\item{dose_col}{Character string specifying the name of the dose column in `data`.}

\item{max_clusters}{Maximum number of clusters to consider. Default is 10.}
}
\value{
A list containing the optimal number of clusters and a dataframe of cluster assignments.
}
\description{
This function determines the optimal number of clusters for dose-response data using the silhouette method.
It returns the optimal number of clusters and the cluster assignments for each gene.
}
\examples{
\dontrun{
  # Assuming 'data' is available
  cluster_results <- optimal_clusters_silhouette(data, dose_col = "Dose")
}
}
