% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_response.R
\name{plot_dotplot_top_pathways}
\alias{plot_dotplot_top_pathways}
\title{Plot Dot Plot of Top Pathways from DoseRider Results}
\usage{
plot_dotplot_top_pathways(
  dose_rider_results,
  top = 10,
  order_column = "NegLogPValue",
  pvalue_column = "best_model_pvalue",
  decreasing = F
)
}
\arguments{
\item{dose_rider_results}{A list containing the results from the DoseRider analysis.}

\item{top}{The number of top pathways to display in the plot. Default is 10.}

\item{order_column}{A character string specifying the column to use for ordering gene sets in the plot.}
}
\value{
A ggplot object representing the dot plot of top pathways.
}
\description{
This function creates a dot plot visualizing the top pathways from DoseRider analysis results based on their adjusted cubic p-value.
The size of the dots represents the number of genes in each pathway, while the color indicates the best model selected for each pathway.
}
\examples{
\dontrun{
  # Assuming dose_rider_results is available
  dot_plot <- plot_dotplot_top_pathways(dose_rider_results, top = 10)
  print(dot_plot)
}

}
