% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_response.R
\name{plot_gene_random_effect_relationship}
\alias{plot_gene_random_effect_relationship}
\title{Plot Scatter Plot for Relationship Between Random Intercepts and Effects in a Gene Set}
\usage{
plot_gene_random_effect_relationship(dose_rider_results, gene_set_name)
}
\arguments{
\item{dose_rider_results}{A list containing the results from the DoseRider analysis.}

\item{gene_set_name}{The name of the gene set for which the plot will be generated.}
}
\value{
A ggplot object representing the scatter plot of random intercepts vs random slopes for the specified gene set.
The plot displays the relationship between random intercepts and random slopes for each gene in the gene set.
}
\description{
This function generates a scatter plot to visualize the relationship between
random intercepts and random slopes for genes within a specified gene set.
It is particularly useful for analyzing the variability and relationship
of gene expressions within a gene set modeled.
}
\examples{
\dontrun{
  # Assuming 'dose_rider_results' contains LMM results for gene sets
  scatter_plot <- plot_gene_random_effect_relationship(dose_rider_results, "Gene Set Name")
  print(scatter_plot)
}

}
