% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_response.R
\name{plot_pathway_response}
\alias{plot_pathway_response}
\title{Plot Pathway Response with Enhanced Visualization at Specific Points}
\usage{
plot_pathway_response(
  dose_rider_results,
  gene_set_name,
  dose_col = "Dose",
  center_values = T
)
}
\arguments{
\item{dose_rider_results}{A list containing the results from the DoseRider analysis.}

\item{gene_set_name}{The name of the gene set for which to plot the response.}

\item{dose_col}{The name of the column representing dose information.}
}
\value{
A ggplot object representing the pathway response plot.
}
\description{
This function creates a plot showing the pathway response with enhanced line thickness
at specific zero points of first and second derivatives.
}
