% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_geneset.R
\name{plot_smooth}
\alias{plot_smooth}
\title{Plot Smooth Curves for Gene Expressions}
\usage{
plot_smooth(result, gene_set_name, dose_col = "dose", center_values = FALSE)
}
\arguments{
\item{result}{A list containing the results from the 'doserider' function.
This list must contain a nested list for each gene set, and each nested list
must have a 'predictions' data frame with columns for 'fit', 'se', and 'gene'.}

\item{gene_set_name}{The name of the gene set to plot.}

\item{center_values}{Logical indicating whether to center the expression values for each gene
around the mean expression across all genes. Defaults to FALSE.}
}
\value{
A ggplot2 object displaying the smooth curves for gene expressions, optionally centered.
}
\description{
This function plots smooth curves for gene expressions along with confidence intervals.
It optionally centers the values around the mean expression for each gene.
}
\examples{
result <- doserider(...)  # result from doserider function
plot_smooth(result, "Gene Set 1", center_values = TRUE)

}
