% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_response.R
\name{plot_top_pathway_responses}
\alias{plot_top_pathway_responses}
\title{Plot Top Significant Pathway Responses}
\usage{
plot_top_pathway_responses(
  dose_rider_results,
  top = 6,
  ncol = 3,
  order_column = "best_model_pvalue",
  decreasing = F,
  dose_col = "Dose",
  center_values = TRUE,
  scale_values = TRUE,
  legend_position = "none",
  text_size = 4,
  margin_space = 0,
  model_metrics = FALSE,
  v_size = 0.5,
  annotate_gene = FALSE,
  annotation_text_size = 5,
  draw_bmd = TRUE,
  plot_original_data = F,
  clusterResults = F
)
}
\arguments{
\item{dose_rider_results}{A list containing the results from doseRider analysis.}

\item{top}{An integer specifying the number of top gene sets to include in the plot. Default is 15.}

\item{order_column}{A character string specifying the column to use for ordering gene sets in the plot.}

\item{scale_values}{Logical, indicating whether to scale the prediction values. Default is TRUE.}

\item{legend_position}{The position of the legend in the plot.}

\item{text_size}{The size of the text in the plot.}

\item{margin_space}{The margin space around the plot.}
}
\value{
A combined ggplot object with top significant pathway response plots.
}
\description{
Creates a combined plot of top significant pathway responses from dose_rider_results.
}
