% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting_helpers.R
\name{prepare_mean_data}
\alias{prepare_mean_data}
\title{Prepare mean data and perform centering/standardization if required}
\usage{
prepare_mean_data(
  gene_set_results,
  dose_col,
  center_values,
  scale_values,
  clusterResults
)
}
\description{
Prepare mean data and perform centering/standardization if required
}
