% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseRider.R
\name{print.DoseRider}
\alias{print.DoseRider}
\title{Print method for DoseRider}
\usage{
\method{print}{DoseRider}(x, ...)
}
\arguments{
\item{x}{A DoseRider object.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
This function prints a summary of a DoseRider object.
It shows the class of the object and the number of gene sets it contains.
}
\examples{
\dontrun{
  # Assuming `dose_rider_result` is a DoseRider object
  print(dose_rider_result)
}

}
