% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseRiderLMM.R
\name{process_gene_set_lmm}
\alias{process_gene_set_lmm}
\title{Process a Single Gene Set for DoseRider Analysis}
\usage{
process_gene_set_lmm(
  se,
  dose_col,
  sample_col,
  omic,
  gmt,
  i,
  minGSsize = 5,
  maxGSsize = 300,
  covariates = c()
)
}
\arguments{
\item{se}{SummarizedExperiment object containing the gene expression data.}

\item{dose_col}{Character string specifying the dose column name in `se`.}

\item{sample_col}{Character string specifying the sample column name in `se`.}

\item{omic}{Character string indicating the type of omics data.}

\item{gmt}{List of gene sets (Gene Matrix Transposed format).}

\item{i}{Index of the current gene set in the `gmt` list.}

\item{minGSsize}{Minimum size of the gene set for analysis (default 5).}

\item{maxGSsize}{Maximum size of the gene set for analysis (default 300).}

\item{covariates}{Vector specifying the covariate column(s) in `se`.}
}
\value{
A list containing analysis results for the gene set, or NULL if the gene set size criteria are not met.
}
\description{
This function processes a single gene set from a gene set collection for DoseRider analysis. It fits
Linear Mixed Models (LMMs) with cubic splines to the gene expression data, calculates model metrics,
significance, Benchmark Dose (BMD), and smoothing values. The results are compiled into a list.
}
\examples{
\dontrun{
# Assuming `se` is a SummarizedExperiment object, `gmt` is a list of gene sets
gene_set_results <- process_gene_set_lmm(se, "dose", "sample", "rnaseq", gmt, 1, 5, 300, "sample")
}

}
