% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd.R
\name{process_single_geneset}
\alias{process_single_geneset}
\title{Process a Single Gene Set for BMD and TCD Statistics}
\usage{
process_single_geneset(
  geneset,
  geneset_name,
  dose_col,
  sample_col,
  covariates,
  omic,
  n_bootstrap,
  ci_level,
  clusterResults,
  z
)
}
\arguments{
\item{geneset}{A single gene set from `dose_rider_results`.}

\item{geneset_name}{The name of the gene set being processed.}

\item{dose_col}{Name of the column representing dose information.}

\item{sample_col}{Name of the column representing sample information.}

\item{covariates}{Vector specifying the covariate column(s).}

\item{omic}{Type of omics data, defaults to "rnaseq".}

\item{n_bootstrap}{The number of bootstrap samples to generate.}

\item{ci_level}{Confidence interval level, default is 0.95.}

\item{clusterResults}{Logical, whether to compute cluster-specific results.}

\item{z}{A parameter passed to the model fitting function.}
}
\value{
A data frame with BMD bounds and TCD cluster statistics for the gene set.
}
\description{
This function processes a single gene set to compute BMD bounds and TCD cluster statistics.
}
