% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{smooth_pathway_trend}
\alias{smooth_pathway_trend}
\title{Generate Smooth Predictions for Pathway Trends}
\usage{
smooth_pathway_trend(
  model,
  long_df,
  dose_col = "dose",
  sample_col = "sample",
  omic = "rnaseq",
  random_effects = TRUE,
  covariates_cols = NULL,
  gene_subset = NULL,
  dose_points = 25,
  sample_subset_size = 10
)
}
\arguments{
\item{model}{A model object, typically of class 'lmerMod' or 'glmerMod'.}

\item{long_df}{A data frame containing the input data with columns corresponding to dose, sample, and other variables.}

\item{dose_col}{The name of the dose variable in `long_df`.}

\item{sample_col}{The name of the sample variable in `long_df`.}

\item{omic}{A character string indicating the type of omics data (e.g., "rnaseq").}

\item{random_effects}{Logical, indicating if random effects (like gene) should be included.}

\item{covariates_cols}{Optional; a vector of names of additional covariates in `long_df`.}
}
\value{
A data frame containing the original data along with the predictions from the model.
}
\description{
This function creates a data frame of smooth predictions based on a provided model. It is useful for
visualizing trends in pathway data, especially in the context of dose-response studies. It handles
different types of omics data and can incorporate covariates.
}
\examples{
\dontrun{
data("mtcars")
model <- lmer(mpg ~ wt + (1|cyl), data = mtcars)
predictions <- smooth_pathway_trend(model = model,
                                   long_df = mtcars,
                                   dose_col = "wt",
                                   sample_col = "cyl",
                                   omic = "rnaseq",
                                   random_effects = TRUE,
                                   covariates_cols = NULL)
}

}
