% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamm.R
\name{smooth_predictions}
\alias{smooth_predictions}
\title{Smooth Predictions Function}
\usage{
smooth_predictions(
  model,
  long_df,
  dose_col = "dose",
  sample_col = "sample",
  covariates_cols = ""
)
}
\arguments{
\item{model}{A GAM object. The model must be of class "gam".}

\item{long_df}{A data frame containing the input data.}

\item{dose_col}{A character string specifying the name of the dose column in the input data frame.}

\item{sample_col}{A character string specifying the name of the sample column in the input data frame.}

\item{covariates_cols}{A character string or a vector of character strings specifying the name(s) of the covariate column(s) in the input data frame.}
}
\value{
A list with the smooth predictions from the GAM model.
}
\description{
This function generates smooth predictions from a given GAM model object.
}
\examples{
\dontrun{
data("mtcars")
model <- mgcv::gam(mpg ~ s(hp), data = mtcars)
predictions <- smooth_predictions(model = model,
                                  dose_col = "hp",
                                  sample_col = "mpg",
                                  covariates_cols = "",
                                  long_df = mtcars)
}

}
