% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{transform_smooth_values}
\alias{transform_smooth_values}
\title{Transform Smooth Values to Gene vs. Dose Dataframe}
\usage{
transform_smooth_values(
  smooth_values,
  dose_col = "Dose",
  expression_col = "predictions",
  center_values = TRUE,
  scale_values = TRUE
)
}
\arguments{
\item{smooth_values}{Dataframe containing smooth values from doseRider results.}

\item{dose_col}{Character string specifying the name of the dose column in `smooth_values`.}

\item{expression_col}{Character string specifying the name of the expression column in `smooth_values`.}

\item{center_values}{Logical, whether to center the gene values (default is TRUE).}

\item{scale_values}{Logical, whether to scale the gene values (default is TRUE).}
}
\value{
A transformed dataframe with genes as rows, doses as columns, and each row centered and scaled.
}
\description{
This function transforms smooth values into a dataframe with genes as rows and doses as columns.
Each row (gene) is centered and scaled (standardized) to normalize the data.
}
\examples{
\dontrun{
  # Assuming 'smooth_values' is available
  transformed_data <- transform_smooth_values(smooth_values, dose_col = "Dose")
}
}
