% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_all_models}
\alias{compare_all_models}
\title{Compare Models Individually Against the Null Model with ANOVA}
\usage{
compare_all_models(
  null_results,
  linear_results = NULL,
  non_linear_fixed_results = NULL,
  non_linear_mixed_results = NULL,
  modelType
)
}
\arguments{
\item{null_results}{A fitted model representing the null or base model.}

\item{linear_results}{A fitted model representing the linear model (optional).}

\item{non_linear_fixed_results}{A fitted model representing the non-linear fixed effects model (optional).}

\item{non_linear_mixed_results}{A fitted model representing the non-linear mixed effects model (optional).}

\item{modelType}{A string indicating the type of model comparison: "LMM" for linear mixed models or "GAMM" for generalized additive mixed models.}
}
\value{
A named list of p-values for the comparisons of each model against the base model (null_results).
If a model is not provided, its p-value is returned as NA.
}
\description{
This function compares the goodness of fit of individual models (linear,
non-linear fixed effects, and non-linear mixed effects) against a null/base model
using ANOVA. P-values are calculated for each model comparison individually.
}
\examples{
# Assuming lmm_null, lmm_linear, lmm_non_linear_fixed, and lmm_non_linear_mixed
# are fitted models:
# compare_models_pvalues <- compare_all_models(lmm_null, lmm_linear, lmm_non_linear_fixed,
# lmm_non_linear_mixed, modelType = "LMM")
}
