% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmd.R
\name{compute_bmd_from_main_trend}
\alias{compute_bmd_from_main_trend}
\title{Compute Benchmark Dose (BMD) from Smoothed Trend}
\usage{
compute_bmd_from_main_trend(smooth_pathway, dose_var, z = 1, center_values = T)
}
\arguments{
\item{smooth_pathway}{Data frame containing smoothed trend predictions, typically from a dose-response model.}

\item{dose_var}{The name of the dose variable in `smooth_pathway`.}

\item{z}{A numeric value specifying the number of standard deviations above the control response
to define the target response for BMD. Default is 1.}
}
\value{
A list or a numeric value representing the calculated BMD. If no BMD is found within a
        reasonable range, NA is returned.
}
\description{
This function calculates the Benchmark Dose (BMD) based on a smoothed trend from model predictions.
The BMD is identified as the dose level where the predicted response exceeds a threshold defined
as a specified number of standard deviations (`z`) above the control response. It's designed to
work with smoothed trend data, typically derived from `lmer` or `glmer` model predictions.
}
