% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseRider.R
\name{DoseRiderParallel}
\alias{DoseRiderParallel}
\title{DoseRiderParallel function}
\usage{
DoseRiderParallel(
  se,
  gmt,
  dose_col = "dose",
  sample_col = "sample",
  covariate = "",
  omic = "rnaseq",
  minGSsize = 5,
  maxGSsize = 300,
  method = "fdr",
  num_cores = 5
)
}
\arguments{
\item{se}{The input SummarizedExperiment object or matrix/data frame with metadata.}

\item{gmt}{The gene set collection as a list with gene sets.}

\item{dose_col}{The name of the column in the metadata representing the dose information.}

\item{sample_col}{The name of the column in the metadata representing the sample information.}

\item{covariate}{The name of the column in the metadata representing the covariate information (optional).}

\item{omic}{The type of omic data used (default is "rnaseq").}

\item{minGSsize}{The minimum gene set size for filtering (default is 5).}

\item{maxGSsize}{The maximum gene set size for filtering (default is 300).}

\item{method}{The p-value adjustment method for FDR correction (default is "fdr").}

\item{num_cores}{The number of cores to use for parallel processing (default is 5).}
}
\value{
A list containing the results of the DoseRider analysis for each gene set.
}
\description{
Perform DoseRider analysis in parallel using multiple cores.
}
\examples{
# Example usage of DoseRiderParallel
results <- DoseRiderParallel(se, gmt, dose_col = "dose", sample_col = "sample", covariate = "age", num_cores = 8)

}
