% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseRider.R
\name{DoseRiderParallel}
\alias{DoseRiderParallel}
\title{Perform DoseRider analysis in parallel using multiple cores.}
\usage{
DoseRiderParallel(
  se,
  gmt,
  dose_col = "dose",
  sample_col = "sample",
  covariates = c(),
  omic = "rnaseq",
  minGSsize = 5,
  maxGSsize = 300,
  method = "fdr",
  num_cores = 5,
  modelType = "LMM",
  clusterResults = TRUE,
  FilterPathway = FALSE,
  pca_threshold = 0.6
)
}
\arguments{
\item{se}{The input SummarizedExperiment object or matrix/data frame with metadata.}

\item{gmt}{The gene set collection as a list with gene sets.}

\item{dose_col}{The name of the column in the metadata representing the dose information.}

\item{sample_col}{The name of the column in the metadata representing the sample information.}

\item{covariates}{The name of the column in the metadata representing the covariate information (optional).}

\item{omic}{The type of omic data used (default is "rnaseq").}

\item{minGSsize}{The minimum gene set size for filtering (default is 5).}

\item{maxGSsize}{The maximum gene set size for filtering (default is 300).}

\item{method}{The p-value adjustment method for FDR correction (default is "fdr").}

\item{num_cores}{The number of cores to use for parallel processing (default is 5).}

\item{modelType}{Type of model to be used for analysis, "LMM" for Linear Mixed Models or "GAMM"
for Generalized Additive Mixed Models. Defaults to "LMM".}

\item{clusterResults}{Boolean, if TRUE the genes within a gene set will be clustered
to find similar expression patterns. Defaults to TRUE.}

\item{FilterPathway}{Boolean, if TRUE the function will apply PCA filtering to detect antagonist patterns. Defaults to FALSE.}

\item{pca_threshold}{Numeric value specifying the variance threshold for PC1 to filter pathways. Default is 0.6.}
}
\value{
A list containing the results of the DoseRider analysis for each gene set.
}
\description{
This function conducts DoseRider analysis in parallel, applying either Linear Mixed Models (LMMs) or Generalized
Additive Mixed Models (GAMMs) to each gene set in the gene matrix transposed (GMT) format. It evaluates dose-response
relationships in gene sets, computing various model metrics and significance.
}
\examples{
\dontrun{
data("SummarizedExperiment")
gmt <- list(geneSet1 = list(genes = c("gene1", "gene2")))
results <- DoseRiderParallel(se, gmt, "dose", "sample", "covariate", "rnaseq", num_cores = 4, modelType = "GAMM")
}

}
