% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation_and_modeling.R
\name{estimate_model_parameters}
\alias{estimate_model_parameters}
\title{Update SummarizedExperiment with Estimated Model Parameters}
\usage{
estimate_model_parameters(se)
}
\arguments{
\item{se}{SummarizedExperiment object containing count data and sample metadata.}
}
\value{
Updated SummarizedExperiment object with size factors, dispersions, and theta values included.
}
\description{
Estimates necessary parameters for DESeq2 or edgeR and updates the SummarizedExperiment object
with size factors, dispersions, and theta values in its rowData and colData.
}
