% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamm.R
\name{fit_gam}
\alias{fit_gam}
\title{Fit Generalized Additive Model (GAM)}
\usage{
fit_gam(formula, data)
}
\arguments{
\item{formula}{A formula for the GAM.}

\item{data}{A data frame containing the data for modeling.}

\item{omic}{A character string specifying the type of omic data.
If 'rnaseq', a negative binomial family is used.}
}
\value{
A GAM model if fitting is successful; NA otherwise.
}
\description{
This function fits a GAM to the provided data using the specified formula.
It adapts to different omics types and handles family specification for RNAseq data.
}
\examples{
\dontrun{
  data("mtcars")
  formula <- mpg ~ s(hp) + s(wt)
  model <- fit_gam(formula, data = mtcars, omic = "base")
}

}
