% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_response.R
\name{plot_gene_set_random_effects}
\alias{plot_gene_set_random_effects}
\title{Plot Ridge Plots for Random Effects in Gene Sets}
\usage{
plot_gene_set_random_effects(
  dose_rider_results,
  dose_col = "Dose",
  top = 10,
  order_column = "best_model_pvalue",
  decreasing = F
)
}
\arguments{
\item{dose_col}{A character string specifying the name of the dose column in the raw expression data.}

\item{top}{An integer specifying the number of top gene sets to include in the plot. Default is 15.}

\item{order_column}{A character string specifying the column to use for ordering gene sets in the plot}
}
\value{
A ggplot object representing the ridge plots of random effects for each gene set.
The plot displays the distribution of random effects for each gene set along the y-axis.
}
\description{
This function generates ridge plots to visualize the distribution of random effects for genes
within each gene set. It is particularly useful for analyzing the variability of gene
expressions within gene sets modeled using Linear Mixed Models (LMMs).
}
\examples{
\dontrun{
  # Assuming 'dose_rider_results' contains LMM results for each gene set
  ridge_plot <- plot_gene_set_random_effects(dose_rider_results)
  print(ridge_plot)
}

}
