% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_response.R
\name{plot_top_pathway_responses}
\alias{plot_top_pathway_responses}
\title{Plot Top Significant Pathway Responses}
\usage{
plot_top_pathway_responses(dose_rider_results, top = 15, ncol = 3)
}
\arguments{
\item{dose_rider_results}{A list containing the results from doseRider analysis.}

\item{top}{The number of top pathways to plot based on significance.}
}
\value{
A combined ggplot object with top significant pathway response plots.
}
\description{
Creates a combined plot of top significant pathway responses from dose_rider_results.
}
