% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseRider.R
\name{process_gene_set}
\alias{process_gene_set}
\title{Perform DoseRider Analysis Using Linear Mixed Models (LMMs) or Generalized Additive Mixed Models (GAMMs)}
\usage{
process_gene_set(
  se,
  dose_col,
  sample_col,
  omic,
  gmt,
  i,
  minGSsize = 5,
  maxGSsize = 300,
  covariates = c(),
  modelType = "LMM"
)
}
\arguments{
\item{se}{SummarizedExperiment object or a matrix/data frame containing gene expression data.}

\item{dose_col}{Name of the column representing dose information.}

\item{sample_col}{Name of the column representing sample information.}

\item{omic}{Type of omics data, defaults to "rnaseq".}

\item{gmt}{List of gene sets, each represented as a list with gene names.}

\item{minGSsize}{Minimum gene set size for analysis, defaults to 5.}

\item{maxGSsize}{Maximum gene set size for analysis, defaults to 300.}

\item{covariates}{Optional, vector specifying the covariate column(s) in `se`.}

\item{modelType}{Type of model to fit for each gene set, options are "LMM" for Linear Mixed Models
and "GAMM" for Generalized Additive Mixed Models. Defaults to "LMM".}

\item{method}{Method for multiple testing adjustment, defaults to "fdr".}
}
\value{
A list containing results for each gene set including various metrics, p-values,
        and adjusted p-values. The structure of results will depend on the model type used.
}
\description{
This function performs DoseRider analysis on gene expression data, applying either Linear Mixed Models (LMMs) or
Generalized Additive Mixed Models (GAMMs) to each gene set defined in the gene matrix transposed (GMT) format.
It evaluates dose-response relationships in the context of gene sets and calculates various model metrics,
significance, Benchmark Dose (BMD), and smoothing predictions.
}
\examples{
\dontrun{
data("SummarizedExperiment")
gmt <- list(geneSet1 = list(genes = c("gene1", "gene2")))
results <- DoseRider(se, gmt, "dose", "sample", "covariate", "rnaseq", modelType = "GAMM")
}

}
