% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doseRider.R
\name{process_gene_set}
\alias{process_gene_set}
\title{Process a single gene set for DoseRider analysis}
\usage{
process_gene_set(se, dose_col, sample_col, omic, gmt, i, minGSsize, maxGSsize)
}
\arguments{
\item{dose_col}{The name of the column representing the dose information in the long_df.}

\item{sample_col}{The name of the column representing the sample information in the long_df.}

\item{gmt}{The gene set collection as a list with gene sets.}

\item{i}{The index of the current gene set in the gmt list.}

\item{minGSsize}{The minimum gene set size for filtering (default is 5).}

\item{maxGSsize}{The maximum gene set size for filtering (default is 300).}

\item{long_df}{The long-format data frame containing the gene expression data for the current gene set.}
}
\value{
A list containing the results of the DoseRider analysis for the current gene set, or NULL if the gene set does not meet the size criteria.
}
\description{
This function processes a single gene set from the gene set collection for DoseRider analysis.
It fits GAM models to the gene expression data and calculates model metrics and significance.
The results are then stored in a list.
}
\examples{
# Example usage of process_gene_set
gene_set_results <- process_gene_set(long_df, dose_col = "dose", sample_col = "sample", gmt, i = 1, minGSsize = 5, maxGSsize = 300)

}
