% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{select_best_model}
\alias{select_best_model}
\title{Select the Best Pathway Model Based on AICc}
\usage{
select_best_model(models)
}
\arguments{
\item{models}{A named list of models, where:
- 'null' corresponds to the model with only a random intercept.
- 'base', 'linear', and 'cubic' are models with varying fixed effects, but all incorporate
  random effects for the genes in the pathways.}
}
\value{
A character string indicating the name of the best model for the pathway.
}
\description{
This function evaluates the goodness-of-fit for models that incorporate random effects
for the genes within pathways. The function uses the AICc criterion to determine the best
model among the provided options. The null model contains only a random intercept for the genes.
}
\details{
The best model is chosen as the one giving the lowest AICc value.
Models with an AICc value not lower than the AICc of the null model minus 2 are eliminated.
}
\examples{
\dontrun{
library(lme4)
data <- data.frame(y = rnorm(100), x = rnorm(100), gene = sample(letters[1:5], 100, replace = TRUE))
null_model <- lmer(y ~ 1 + (1|gene), data = data)
base_model <- lmer(y ~ x + (1|gene), data = data)
linear_model <- lmer(y ~ poly(x, 2) + (1|gene), data = data)
cubic_model <- lmer(y ~ poly(x, 3) + (1|gene), data = data)
models_list <- list(null = null_model, base = base_model, linear = linear_model, cubic = cubic_model)
best <- select_best_model(models_list)
print(best)
}
}
